Moved show-jobs-component to the show-jobs folder and renamed it to follow naming conventions
