+++ meteor-app/imports/ui/pages/show-jobs/show-job-component.tsx
